+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * new function `copy_labels_from()`, compliant with `dplyr` syntax
 * `update_labelled()` is now more strict (#42 by @iago-pssjd)
+* new function `lookfor()` imported from `questionr` (#44)
 
 # labelled 2.0.1